department,programs,enroll,student,works,org. table completed

+++ controllers/StudentOrganizationController.php
@@ -56,9 +56,14 @@ class StudentOrganizationController extends Controller
      */
     public function actionView($id)
     {
+        if(!Yii::$app->user->isGuest){
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }else {
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -69,6 +74,7 @@ class StudentOrganizationController extends Controller
     public function actionCreate()
     {
         $model = new StudentOrganization();
+        if(!Yii::$app->user->isGuest){
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->student_organization_id]);
@@ -77,6 +83,9 @@ class StudentOrganizationController extends Controller
         return $this->render('create', [
             'model' => $model,
         ]);
+        }else {
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -89,6 +98,7 @@ class StudentOrganizationController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
+        if(!Yii::$app->user->isGuest){
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->student_organization_id]);
@@ -97,6 +107,9 @@ class StudentOrganizationController extends Controller
         return $this->render('update', [
             'model' => $model,
         ]);
+        }else {
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -108,9 +121,14 @@ class StudentOrganizationController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->isGuest){
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }else {
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
